requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
pandas
Pillow
tqdm
scipy
scikit-learn
tk
